Add files via upload
Chatbot model designed to perform costumer service for a dental clinic, which is able to interact with the user, offer information about the services they provide, answer specific requests and schedule appointments. This chatbot has been designed to interact in Spanish, as the clinic it is oriented to is based in Spain.