Add files via upload
Chatbot model designed to perform costumer service for a dental clinic based in Barcelona, Spain. These chatbot is oriented towards answering basic questions regarding the location and services offered by the clinic. It is designed to work in the Spanish language, as it is one of the most languages spoken in this location. This is a first version of this chatbot, it is open for changes and improvement to perform a better service for the user.